+++ ControlVeterinariaAPI/src/main/java/org/esfe/DTOS/veterinario/VeterinarioModificar.java
@@ -1,7 +1,6 @@
 package org.esfe.DTOS.veterinario;
 
 import lombok.*;
-
 import java.time.LocalDate;
 
 @Data
@@ -9,10 +8,10 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @Builder
 public class VeterinarioModificar {
-    private Short id;
-    private Byte especializacionId;
+    private int id;
+    private int especializacionId;
     private String numeroLicencia;
-    private Byte rolId;
+    private int rolId;
     private String nickName;
     private String correo;
     private String nombreCompleto;
